[iOS] refactor: SectionAmenities 개수 제한 구현 변경
- 원래 데이터를 보존하도록 변경
- SectionMap 관련 코드 별도 파일로 분리

issue: #108
