+++ iOS/AirbnbApp/AirbnbApp/StayDetail/Sections/ViewFactories/SectionAmenitiesFactory.swift
@@ -2,19 +2,21 @@ import Foundation
 
 class SectionAmenitiesFactory {
     static func makeView(for stayDetail: StayDetail, amenitiesLimit number: Int) -> SectionAmenities {
-        var amenities = AmenityCollection(for: stayDetail)
-        return makeView(for: amenities.limited(number))
+        let amenities = AmenityCollection(for: stayDetail)
+        return makeView(for: amenities, amenitiesLimit: number)
     }
 
-    static func makeView(for amenities: AmenityCollection) -> SectionAmenities {
+    static func makeView(for amenities: AmenityCollection, amenitiesLimit number: Int) -> SectionAmenities {
         let view = SectionAmenities.loadFromXib()
+        view.viewAllButton.setTitle("Show all \(amenities.count) amenities", for: .normal)
 
-        amenities.items.map {
-            let view = AmenityView.loadFromXib()
-            view.titleLabel.text = $0.title
-            view.symbolImageView.image = $0.symbol
+        let limit = min(number, amenities.count)
+        amenities.items[0..<limit].map {
+            let amenityView = AmenityView.loadFromXib()
+            amenityView.titleLabel.text = $0.title
+            amenityView.symbolImageView.image = $0.symbol
 
-            return view
+            return amenityView
         }.forEach { view.stackView.addArrangedSubview($0) }
 
         return view
