+++ iOS/AirbnbApp/AirbnbApp/StayDetail/Sections/ViewFactories/SectionSummaryFactory.swift
@@ -1,1 +1,11 @@
 import Foundation
+
+class SectionSummaryFactory {
+    static func makeView(for stayDetail: StayDetail) -> SectionSummary {
+        let view = SectionSummary.loadFromXib()
+
+        view.textLabel.text = stayDetail.roomInfo.summary
+
+        return view
+    }
+}
