+++ iOS/AirbnbApp/AirbnbApp/StayDetail/Sections/Views/SectionSummary.xib
@@ -10,42 +10,40 @@
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
         <view contentMode="scaleToFill" id="iN0-l3-epB" customClass="SectionSummary" customModule="AirbnbApp" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="414" height="140"/>
+            <rect key="frame" x="0.0" y="0.0" width="414" height="104"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
-                <textView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" scrollEnabled="NO" showsHorizontalScrollIndicator="NO" editable="NO" textAlignment="natural" translatesAutoresizingMaskIntoConstraints="NO" id="mai-vb-8QM">
-                    <rect key="frame" x="0.0" y="0.0" width="414" height="100"/>
-                    <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
-                    <constraints>
-                        <constraint firstAttribute="height" relation="lessThanOrEqual" constant="100" id="IrN-fo-Loq"/>
-                    </constraints>
-                    <string key="text">Lorem ipsum dolor sit er elit lamet, consectetaur cillium adipisicing pecu, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. Nam liber te conscient to factor tum poen legum odioque civiuda.</string>
-                    <color key="textColor" systemColor="labelColor" cocoaTouchSystemColor="darkTextColor"/>
-                    <fontDescription key="fontDescription" type="system" pointSize="14"/>
-                    <textInputTraits key="textInputTraits" autocapitalizationType="sentences"/>
-                </textView>
                 <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="aKY-Ju-gSm">
-                    <rect key="frame" x="0.0" y="110" width="49" height="30"/>
+                    <rect key="frame" x="0.0" y="74" width="49" height="30"/>
                     <state key="normal" title="more &gt;">
                         <color key="titleColor" cocoaTouchSystemColor="darkTextColor"/>
                     </state>
                 </button>
+                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="249" verticalCompressionResistancePriority="749" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="3" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="BuG-lQ-b8N">
+                    <rect key="frame" x="0.0" y="0.0" width="414" height="64"/>
+                    <string key="text">Label
+Label
+Label</string>
+                    <fontDescription key="fontDescription" type="system" pointSize="14"/>
+                    <nil key="textColor"/>
+                    <nil key="highlightedColor"/>
+                </label>
             </subviews>
             <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
             <constraints>
-                <constraint firstItem="mai-vb-8QM" firstAttribute="trailing" secondItem="iN0-l3-epB" secondAttribute="trailing" id="0og-9n-SQ7"/>
-                <constraint firstItem="mai-vb-8QM" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" id="87T-7t-Fqd"/>
-                <constraint firstItem="aKY-Ju-gSm" firstAttribute="top" secondItem="mai-vb-8QM" secondAttribute="bottom" constant="10" id="hcs-rP-j2i"/>
+                <constraint firstItem="BuG-lQ-b8N" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" id="1oA-6h-Qv3"/>
+                <constraint firstItem="BuG-lQ-b8N" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" id="Cs4-CN-AuB"/>
+                <constraint firstItem="aKY-Ju-gSm" firstAttribute="top" secondItem="BuG-lQ-b8N" secondAttribute="bottom" constant="10" id="ISV-Xr-z9o"/>
+                <constraint firstItem="BuG-lQ-b8N" firstAttribute="trailing" secondItem="vUN-kp-3ea" secondAttribute="trailing" id="UWr-a1-H2Q"/>
                 <constraint firstItem="aKY-Ju-gSm" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" id="mEq-bk-Y1H"/>
                 <constraint firstAttribute="bottom" secondItem="aKY-Ju-gSm" secondAttribute="bottom" id="mmW-PL-6Dk"/>
-                <constraint firstItem="mai-vb-8QM" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" id="qPc-NX-1XJ"/>
             </constraints>
             <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
             <viewLayoutGuide key="safeArea" id="vUN-kp-3ea"/>
             <connections>
-                <outlet property="textView" destination="mai-vb-8QM" id="Q5f-3K-ZCg"/>
+                <outlet property="textLabel" destination="BuG-lQ-b8N" id="2B0-yr-6fy"/>
             </connections>
-            <point key="canvasLocation" x="-333.33333333333337" y="-567.1875"/>
+            <point key="canvasLocation" x="-333.33333333333337" y="-579.91071428571422"/>
         </view>
     </objects>
 </document>
